requirements.txt
boto3==1.28.59
botocore==1.31.85
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.8.0
numpy==1.26.4
packaging==24.2
pandas==2.1.2
pillow==11.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
ruptures==1.1.9
s3transfer==0.7.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.17.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.0.7
